# Reduce optimization on files with long build times on Chrysalis